rename core service file to remove the s
